refactor(profile): split toml_store into directory module
Part of #290
