+++ src/crosshook-native/crates/crosshook-core/src/profile/toml_store/tests/mod.rs
@@ -1,0 +1,8 @@
+mod fixtures;
+
+mod duplicate;
+mod import_legacy;
+mod load_save;
+mod manual_presets;
+mod optimizations;
+mod shareable_toml;
